Adding a WIP role for managing Azure infrastructure for OpenShift using resource manager templates. (#170)
An important thing to note is the template in the templates directory is not being directly used, but the same contents (translated to yaml) are used in the generate-templates.yml parameters for the 'Create Azure Deploy' task. The template will ideally be referenced directly from the Github url once it is complete.